# Low-allocation version of JSON Pointer and other improvements