--- a/JsonPatch.Tests/JsonPatch.Tests.csproj
+++ b/JsonPatch.Tests/JsonPatch.Tests.csproj
@@ -21,6 +21,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\JsonPatch\JsonPatch.csproj" />
+		<ProjectReference Include="..\JsonSchema\JsonSchema.csproj" />
 	</ItemGroup>
 
 </Project>