--- a/JsonPatch.Tests/Suite/JsonPatchTest.cs
+++ b/JsonPatch.Tests/Suite/JsonPatchTest.cs
@@ -3,7 +3,7 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
 using Json.More;
-using Json.Schema;
+//using Json.Schema;
 // ReSharper disable ConditionIsAlwaysTrueOrFalse
 #pragma warning disable CS8618
 
@@ -12,66 +12,66 @@ namespace Json.Patch.Tests.Suite;
 [JsonConverter(typeof(JsonPatchTestJsonConverter))]
 public class JsonPatchTest
 {
-	public static readonly JsonSchema TestSchema = new JsonSchemaBuilder()
-		.Schema(MetaSchemas.Draft202012Id)
-		.Defs(
-			("operationType", new JsonSchemaBuilder().Enum(
-					"add",
-					"remove",
-					"replace",
-					"move",
-					"copy",
-					"test"
-				)
-			)
-		)
-		.Type(SchemaValueType.Object)
-		.Properties(
-			("doc", true),
-			("expected", true),
-			("patch", new JsonSchemaBuilder()
-				.Type(SchemaValueType.Array)
-				.Items(new JsonSchemaBuilder()
-					.Type(SchemaValueType.Object)
-					.Properties(
-						("op", new JsonSchemaBuilder().Ref("#/$defs/operationType")),
-						("path", new JsonSchemaBuilder()
-							.Type(SchemaValueType.String)
-							.Format(Formats.JsonPointer)
-						),
-						("from", new JsonSchemaBuilder()
-							.Type(SchemaValueType.String)
-							.Format(Formats.JsonPointer)
-						),
-						("value", true)
-					)
-					.Required("op")
-					.OneOf(
-						new JsonSchemaBuilder()
-							.Properties(("op", new JsonSchemaBuilder().Const("add")))
-							.Required("path", "value"),
-						new JsonSchemaBuilder()
-							.Properties(("op", new JsonSchemaBuilder().Const("remove")))
-							.Required("path"),
-						new JsonSchemaBuilder()
-							.Properties(("op", new JsonSchemaBuilder().Const("replace")))
-							.Required("path", "value"),
-						new JsonSchemaBuilder()
-							.Properties(("op", new JsonSchemaBuilder().Const("move")))
-							.Required("path", "from"),
-						new JsonSchemaBuilder()
-							.Properties(("op", new JsonSchemaBuilder().Const("copy")))
-							.Required("path", "from"),
-						new JsonSchemaBuilder()
-							.Properties(("op", new JsonSchemaBuilder().Const("test")))
-							.Required("path", "value")
-					)
-				)
-			),
-			("comment", new JsonSchemaBuilder().Type(SchemaValueType.String)),
-			("error", new JsonSchemaBuilder().Type(SchemaValueType.String)),
-			("disabled", new JsonSchemaBuilder().Type(SchemaValueType.Boolean))
-		);
+	//public static readonly JsonSchema TestSchema = new JsonSchemaBuilder()
+	//	.Schema(MetaSchemas.Draft202012Id)
+	//	.Defs(
+	//		("operationType", new JsonSchemaBuilder().Enum(
+	//				"add",
+	//				"remove",
+	//				"replace",
+	//				"move",
+	//				"copy",
+	//				"test"
+	//			)
+	//		)
+	//	)
+	//	.Type(SchemaValueType.Object)
+	//	.Properties(
+	//		("doc", true),
+	//		("expected", true),
+	//		("patch", new JsonSchemaBuilder()
+	//			.Type(SchemaValueType.Array)
+	//			.Items(new JsonSchemaBuilder()
+	//				.Type(SchemaValueType.Object)
+	//				.Properties(
+	//					("op", new JsonSchemaBuilder().Ref("#/$defs/operationType")),
+	//					("path", new JsonSchemaBuilder()
+	//						.Type(SchemaValueType.String)
+	//						.Format(Formats.JsonPointer)
+	//					),
+	//					("from", new JsonSchemaBuilder()
+	//						.Type(SchemaValueType.String)
+	//						.Format(Formats.JsonPointer)
+	//					),
+	//					("value", true)
+	//				)
+	//				.Required("op")
+	//				.OneOf(
+	//					new JsonSchemaBuilder()
+	//						.Properties(("op", new JsonSchemaBuilder().Const("add")))
+	//						.Required("path", "value"),
+	//					new JsonSchemaBuilder()
+	//						.Properties(("op", new JsonSchemaBuilder().Const("remove")))
+	//						.Required("path"),
+	//					new JsonSchemaBuilder()
+	//						.Properties(("op", new JsonSchemaBuilder().Const("replace")))
+	//						.Required("path", "value"),
+	//					new JsonSchemaBuilder()
+	//						.Properties(("op", new JsonSchemaBuilder().Const("move")))
+	//						.Required("path", "from"),
+	//					new JsonSchemaBuilder()
+	//						.Properties(("op", new JsonSchemaBuilder().Const("copy")))
+	//						.Required("path", "from"),
+	//					new JsonSchemaBuilder()
+	//						.Properties(("op", new JsonSchemaBuilder().Const("test")))
+	//						.Required("path", "value")
+	//				)
+	//			)
+	//		),
+	//		("comment", new JsonSchemaBuilder().Type(SchemaValueType.String)),
+	//		("error", new JsonSchemaBuilder().Type(SchemaValueType.String)),
+	//		("disabled", new JsonSchemaBuilder().Type(SchemaValueType.Boolean))
+	//	);
 
 	public JsonNode? Doc { get; set; }
 	public JsonNode? ExpectedValue { get; set; }
@@ -108,13 +108,13 @@ public class JsonPatchTestJsonConverter : WeaklyTypedJsonConverter<JsonPatchTest
 
 		try
 		{
-			var results = JsonPatchTest.TestSchema.Evaluate(node, new EvaluationOptions
-			{
-				OutputFormat = OutputFormat.Hierarchical,
-				RequireFormatValidation = true
-			});
-			if (results.IsValid)
-			{
+			//var results = JsonPatchTest.TestSchema.Evaluate(node, new EvaluationOptions
+			//{
+			//	OutputFormat = OutputFormat.Hierarchical,
+			//	RequireFormatValidation = true
+			//});
+			//if (results.IsValid)
+			//{
 				var model = node.Deserialize<Model>(options)!;
 				return new JsonPatchTest
 				{
@@ -126,9 +126,9 @@ public class JsonPatchTestJsonConverter : WeaklyTypedJsonConverter<JsonPatchTest
 					Patch = model.Patch,
 					Disabled = model.Disabled
 				};
-			}
+			//}
 
-			Console.WriteLine(JsonSerializer.Serialize(results, TestEnvironment.SerializerOptions));
+			//Console.WriteLine(JsonSerializer.Serialize(results, TestEnvironment.SerializerOptions));
 			return null;
 		}
 		catch (Exception e) when(e is JsonException or ArgumentException)