… Benchmarks/Benchmarks.csproj
… Benchmarks/Pointer/Runner.cs
… Benchmarks/Program.cs
… Benchmarks/SchemaSuite/TestSuiteRunner.cs
… JsonPatch/EnumerableExtensions.cs
… JsonPatch/JsonPatch.cs
… JsonPatch/JsonPatch.csproj
… JsonPatch/MoveOperationHandler.cs
… JsonPatch/PatchExtensions.cs
… JsonPatch/PatchOperation.cs
… JsonPatch/PointerExtensions.cs
… JsonPatch/RemoveOperationHandler.cs
… JsonPatch/ReplaceOperationHandler.cs
… JsonPointer.Tests/JsonPointerParseTests.cs
… JsonPointer.Tests/ManipulationTests.cs
… JsonPointer.Tests/RelativeJsonPointerParseTests.cs
… JsonPointer.Tests/SpecificationTests.cs
… JsonPointer.Tests/TypeConverterTests.cs
… JsonPointer/EnumerableExtensions.cs
… JsonPointer/JsonPointer.cs
… JsonPointer/JsonPointer.csproj
… JsonPointer/JsonPointerJsonConverter.cs
… JsonPointer/JsonPointerStyle.cs
… JsonPointer/JsonPointerTypeConverter.cs
… JsonPointer/NodeExtensions.cs
… JsonPointer/PointerCreationOptions.cs
… JsonPointer/PointerSegment.cs
… JsonPointer/RelativeJsonPointer.cs
… JsonPointer/SpanExtensions.cs
… JsonSchema.ArrayExt/JsonSchema.ArrayExt.csproj