Implements a low-allocation version of JSON Pointer.

API is _mostly_ the same, but it does break.  Primarily, `JsonPointer` is now a struct.

Previously, `JsonPointer` held a collection of `PointerSegment`s.  These were all classes and each segment held a string.

Now, `JsonPointer` is a struct that holds a string representation of the pointer and a collection of `Range`s that identify each segment.  This architectural change improved run time and memory consumption a little.

Further investigation revealed that much of the memory allocation came from actually parsing the string, where lists were used to manage collections.  Because these lists are only used within the method scope, we can make another optimization to use `Memory<Range>` instead (via the `MemoryPool<T>` class).  This change cut run time by 40% and memory usage by 51%!

## Parsing Pointers

This benchmark parses the pointers that are in the spec _n_ times.

| Method | n | Mean       | Error     | StdDev    | Gen0     | Allocated |
|------- |------ |-----------:|----------:|----------:|---------:|----------:|
| master branch | 1     |   5.830 us |  0.2559 us |  0.7300 us |   4.6387 |   9.48 KB |
| local  | 1     |   3.188 us | 0.0492 us | 0.0460 us |   2.2583 |   4.62 KB |
| master branch| 10    |  54.627 us |  1.0469 us |  2.9355 us |  46.3867 |  94.77 KB |
| local  | 10    |  31.482 us | 0.4489 us | 0.4199 us |  22.5830 |  46.17 KB |
| master branch| 100   | 537.704 us | 10.6843 us | 22.5369 us | 463.8672 | 947.66 KB |
| local  | 100   | 329.112 us | 2.6580 us | 2.4863 us | 226.0742 | 461.72 KB |

## JSON Schema Test Suite

Additional improvements were made to schema processing, including refactoring to perform more analysis at registration time and more targeted memory management.

This benchmark runs the test suite _n_ times.

| Method          | n  | Mean       | Error    | StdDev   | Gen0        | Gen1       | Allocated |
|---------------- |--- |-----------:|---------:|---------:|------------:|-----------:|----------:|
| master branch   | 1  |   412.7 ms | 14.16 ms | 41.30 ms |  27000.0000 |  1000.0000 |  82.66 MB |
| current         | 1  | 301.6 ms   |  5.93 ms | 10.07 ms |  23000.0000 | 7000.0000  |  78.41 MB |
| master branch   | 10 | 1,074.7 ms | 22.24 ms | 63.82 ms | 218000.0000 | 11000.0000 | 476.56 MB |
| current         | 10 | 945.9 ms   | 18.64 ms | 32.15 ms | 216000.0000 | 5000.0000  | 472.94 MB |

(more improvements below!)