@gregsdennis 2024-04-30
Result of last commit

## parse

| Method | Count | Mean       | Error     | StdDev    | Gen0     | Allocated |
|------- |------ |-----------:|----------:|----------:|---------:|----------:|
| 4.0.1  | 1     |   2.941 us | 0.0573 us | 0.0637 us |   4.1962 |   8.57 KB |
| 5.0.0  | 1     |   3.825 us | 0.0760 us | 0.0961 us |   3.0823 |    6.3 KB |
| 4.0.1  | 10    |  27.367 us | 0.5281 us | 0.6287 us |  41.9617 |   85.7 KB |
| 5.0.0  | 10    |  36.155 us | 0.6979 us | 0.9074 us |  30.8228 |  62.97 KB |
| 4.0.1  | 100   | 298.443 us | 4.6407 us | 4.1139 us | 419.4336 | 857.03 KB |
| 5.0.0  | 100   | 362.064 us | 6.7056 us | 6.2724 us | 308.1055 | 629.69 KB |

## combine

| Method | Count | Mean        | Error     | StdDev    | Gen0     | Allocated |
|------- |------ |------------:|----------:|----------:|---------:|----------:|
| 4.0.1  | 1     |    679.3 ns |  12.35 ns |  10.95 ns |   1.1473 |   2.34 KB |
| 5.0.0  | 1     |    538.2 ns |  10.12 ns |  10.83 ns |  0.9794 |      2 KB |
| 4.0.1  | 10    |  6,493.4 ns | 127.92 ns | 125.63 ns |  11.4746 |  23.44 KB |
| 5.0.0  | 10    |  5,188.1 ns |  97.80 ns | 104.65 ns |  9.7885 |     20 KB |
| 4.0.1  | 100   | 65,566.7 ns | 682.27 ns | 569.73 ns | 114.7461 | 234.38 KB |
| 5.0.0  | 100   | 58,245.0 ns | 646.43 ns | 539.80 ns | 97.9004 |    200 KB |

## schema

| Version  | n  | Mean       | Error    | StdDev   | Gen0        | Gen1       | Allocated |
|----------|--- |-----------:|---------:|---------:|------------:|-----------:|----------:|
| v6.1.0   | 1  |   412.7 ms | 14.16 ms | 41.30 ms |  27000.0000 |  1000.0000 |  82.66 MB |
| v7.0.0   | 1  | 299.1 ms |  5.82 ms |  9.06 ms |  25000.0000 | 5000.0000 |  74.33 MB |
| v6.1.0   | 10 | 1,074.7 ms | 22.24 ms | 63.82 ms | 218000.0000 | 11000.0000 | 476.56 MB |
| v7.0.0   | 10 | 893.2 ms | 17.64 ms | 38.71 ms | 199000.0000 | 6000.0000 | 436.48 MB |