@gregsdennis 2024-04-30
After consolidating keyword metadata into keyword registry:

| Version  | n  | Mean     | Error    | StdDev   | Gen0        | Gen1      | Allocated |
|----------|--- |---------:|---------:|---------:|------------:|----------:|----------:|
| v7.0.0   | 1  | 296.5 ms |  5.82 ms | 10.03 ms |  21000.0000 | 4000.0000 |  72.81 MB |
| v7.0.0   | 10 | 903.0 ms | 17.96 ms | 40.91 ms | 202000.0000 | 9000.0000 | 443.65 MB |